Debugging a while loop is more accurate
Previously when using the debugger, the while loop would not break on
the final test of a while loop but instead jump from the last statement
to the after while block. The debug block is now positioned so that the
last test that causes the loop to end is shown by the debugger.